Rename identify.py to pem.py.
